Former-commit-id: 954216e49d78c45b0f66ef7d83aff966f5aa7fdf

+++ modules/SofaBaseCollision/SphereModel.inl
@@ -1,6 +1,6 @@
 /******************************************************************************
 *       SOFA, Simulation Open-Framework Architecture, development version     *
-*                (c) 2006-2015 INRIA, USTL, UJF, CNRS, MGH                    *
+*                (c) 2006-2016 INRIA, USTL, UJF, CNRS, MGH                    *
 *                                                                             *
 * This library is free software; you can redistribute it and/or modify it     *
 * under the terms of the GNU Lesser General Public License as published by    *
@@ -112,7 +112,7 @@ void TSphereModel<DataTypes>::draw(const core::visual::VisualParams* vparams,int
 {
     TSphere<DataTypes> t(this,index);
 
-	vparams->drawTool()->drawSphere(t.p(), t.r());
+    vparams->drawTool()->drawSphere(t.p(), (float)t.r());
 }
